requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.1
Django==5.2.5
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-filter==25.1
django-js-asset==3.1.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
firebase_admin==7.1.0
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
http_ece==1.2.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
kombu==5.5.4
lxml==6.0.1
msgpack==1.1.1
multidict==6.6.4
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.1.1
pytz==2025.2
pywebpush==2.0.3
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
